+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [x] Search details and availability
 - [x] Update libraries opening hours
+- [x] Add English localization in Library screen
+- [x] Improve Dynamic Type support in Library screen
 
 
 ## 1.6