+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # BM Grenoble
 
+## 1.6.2 (work in progress)
+
+
 ## 1.6.1
 
 2021-03-25